refactor: create exception dir
