+++ src/fukinotou/image_loader.py
@@ -4,7 +4,7 @@ from typing import List
 from pydantic import BaseModel, ConfigDict
 from PIL import Image
 
-from .load_error import LoadingError
+from .exception.loading_error import LoadingError
 from .path_handler.path_searcher import PathSearcher
