+++ src/fukinotou/parquet_loader.py
@@ -5,7 +5,7 @@ from typing import List, Type, TypeVar, Generic
 from pydantic import BaseModel, ValidationError
 
 from fukinotou.dataframe_exportable import DataframeExportable
-from fukinotou.load_error import LoadingError
+from fukinotou.exception.loading_error import LoadingError
 
 T = TypeVar("T", bound=BaseModel)
